fix(#97): update import paths after extracting http_server from the internal package to make it importable.
